+++ CHANGELOG.md
@@ -8,6 +8,7 @@ A persistence layer for Hanami
 ### Fixed
 - [Marcello Rocha] Fixed migrations MySQL detection of username and password
 - [Luca Guidi] Fixed migrations creation/drop of a MySQL database with a dash in the name
+- [Semyon Pupkov] Ensure `db console` to work when Postgres connection URL is defined with `"postgresql://"` scheme
 
 ## v1.0.0.beta1 - 2017-02-14
 ### Added